Vidjil: Accelerate KmerRepresentativeComputer
Using an ArrayKmerStore was definitely a bad idea. For every clone it was creating each time a new array
of 4^14 elements. And each element was initalized with an empty Kmer().
Rather we use a MapKmerStore, where only the used elements are defined, which is therefore quicker
(even if accesses to MapKmerStore are in fact slower).