Interface: Bug when the json input file has a V that is not known by the interface.
Vidjil CLI and Vidjil interface are not using the same files for the
repertoire sequences. So, that's natural this bug happens. Could the JS file
parse directly the FASTA files instead of duplicating the repertoires in
germline.js?

With this solution there could still have an issue when loading an old data
file that contains segmentations with V (or J or D) sequences that have been
removed from the repertoire. Therefore the interace must be able to deal with
unknown sequences. Should it fallback to undefined? Something else?