segmenter.cpp: keeps only the first sequence when there are several best matching sequences
When the best alignment score is reached by several germline sequences, a choice has to be
made on the sequence to be kept. This choice was arbitrarily set to the first processed sequence.

However, previously, this choice was the case for all alignment informations (deletions...)
except for the name of the germline sequence, which was retrieven from the score list.
This list is sorted by a non-stable sort, thus the name of the sequence was somewhat random.

This is corrected now, and the list is used only in the .json output (to be used later in the interface).

- Closes bug 2013-12-18.
- By chance, all other should_get tests on segmentation are still valid