Vidjil: Catch error messages from git.
If git is not available of if we're not in a git repo, we don't want the error
message to be displayed. In that case we just say that we're not in a git
repository.

system() returns -1 on error (can't fork) or the return value of the
program. Hence, if git returns correctly system returns 0. In every other
cases, it is a problem.