KmerSegmenter: Cleaned KmerSegmenter.
KmerSegmenter doesn't fill a stats array anymore. It doesn't need anymore a
Cost. It doesn't write anymore the unsegmented sequences in a stream.

Instead a function gives access to the reason of segmentation or
unsegmentation. It also provides an access to the KmerAffectAnalyser of the
sequence. So all the informations are available from the outside of the class
and we can use it to write or count whatever we want.
This will be done in another class.