AffectAnalyser: Provide a constant-time counting affectation solution
We count affectations in constant-time before or after a position. This will
help improving the heuristic for finding the window

firstMax and lastMax are the most useful functions since they provide the
position of the first (resp. last) positions where the maximum is achieved for
some affectations. So that gives us the range where we should center our
window.