Affect: Bug correction regarding different strands with Kmer(String)Affect
When a k-mer, for the same label, is seen on both strand, we were (actually, I
was) putting the + strand by default. But that's not robust to revcomp
sequences. That explains some different results on strand + and strand - on
very specific cases (mainly with small k).

Results from test stanford-w100 is updated accordingly.