Tests: S22 with revcomp.
Instead of segmenting 75% of the reads we now segment 99.9%.
Therefore the hierarchy among clones is slightly different and the clone which was the first is now second.
Additionnally the bug correction of the strand in KmerAffect as solved two problems regarding the number
of reads segmented.

Still there is a little problem with the number of windows segmented. I guess that has something to do
with the rounding of the position of the center of the window