Tests: Less windows found in Stanford S22 with -w 100
No worries this comes from the fact that this tests uses a very large
window. Therefore a slight modification of the segmentation may shift the
segmentation to the right and may prevent from getting a window.

There were several cases with an affectation like that one:
+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V ?+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V ?+V+V+V+V+V+V+V+V ? ? ?+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V ?+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V ?+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V+V _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _+V _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _+V _ _ _ _ _ _ _ _ _+J+J ?+J+J+J+J+J+J+J+J+J+J+J+J+J+J+J+J+J+J+J+J+J+J+J ?+J+J+J+J

Where the last +V was not taken into account with the previous heuristic, but
was with the current one.