vidjil.cpp: output 'out/segmented.vdj.fa' only when the new option '-U' is set.
A very long time ago, 'out/segmented.vdj.fa' was considered to be
the main output of Vidjil, giving the (heuristic) segmentation of
each read. However, when many reads are segmented, writing this file
to the disk can slow the software (until 25%)... and this file
was used only to debug. By default, this file is now not produced.

The main output, with a summary of windows and representative, is in
'out/windows.fa', 'out/seq/clone.fa-' and now 'out/vidjil.data'.