vidjil.cpp: large simplification, do no consider the clusters as main elements
The clusters computed by cluster-junctions.cpp are meant to be visualized in the browser,
and do not influence anymore the main output of vidjil.cpp: one clone = one window.
This brings a great simplification in the code: a clone is now a 'junction' and not a 'list <junction>'.

Now 'detailed_cluster_analysis' only controls whether a representative is computed or not.