fuse.py: ignore non-list values in concatenate_with_padding()
This was introduced by b9cd02c, but removed with b9cd02c. This prevents fuse.py from halting
if unknown non-list fields are present.

I am not sure that this is useful (and here it's somewhat different: such values are ignored).
We need more unit and functional tests on fuse.py !