core/segment.cpp: UNSEG_AMBIGUOUS for non-segmented sequences that have many V and J.
When a sequence has both enough V and J affectations (above AMBIGUOUS_THRESHOLD),
it should normally be segmented. If not, the sequence is probably a chimera,
or the V and J repertoires are not well choosen. We now label this sequence
as AMBIGUOUS (instead of TOO_FEW_V or TOO_FEW_J).