core/segment.cpp: UNSEG_AMBIGUOUS of the previous commits becomes a 'detected' flag
ABIGUOUS_THRESHOLD becomes DETECTED_THRESHOLD, such reads are flag 'detected'

This is cleaner: on one side we have the cause of unsegmentation, on the other we have the
flag 'detected' if there were enough V+J (now total V+J, we don't need to have both V and J).
Currently, the flag can be set when SEG_+/, TOO_FEW_V/J, or DELTA_MIN/MAX (but not STRAND).

We do not continue to the remaining germlines when 'detected' or STRAND.