core/segment.{h,cpp}: both V and J kmers must be above DETECT_THRESHOLD to flag the germline as 'detected'
This reverts the idea of 41b4f0710, but still uses a 'detected' flag.
This flag tells us whether we should continue to the remaining germlines.
In the case of incomplete rearrangements, we certainly want to continue even if we have
many V or many J (but now both of them).
Tests d2a0dbd58c are now passing.

Note that we don't run this test when strand == 2, this should be changed.