test_browser: HTML elements are now accessed through VidjilBrowser
Class VidjilBrowser knows how to get the elements in the browser (it knows the class, the hierarchy and so on).
Inside the tests we should rely as much as possible on VidjilBrowser rather than using directly class names, ID or texts
to retrieve a given element.

Of course VidjilBrowser should be completed all along the time :)