core/segment.cpp: KmerMultiSegmenter test all KmerSegmenter, and returns the best one
This is an important change in how the heuristics works with multi-germlines.

Now the segmentation may run 2x to 3x slower (with -g -i), but we ensure that
the best germline is selected for each read. This improves the results,
in particular with mixed data from regular and incomplete rearrangements.