MultiKmerSegmenter: Use pointer for the_kseg
Not using a pointer creates some problems with KmerSegmenter.
We should have a copy constructor in KmerSegmenter but it is a pain in the ***.

Not using pointers makes several copies of the same object. However the copies share the
same pointer to a KmerAffectAnalyser and that's the problem. When one of the copy is
deleted, its destructor is called which will delete the KmerAffectAnalyser and the other
copies won't have access to it, and it will also generate an error when the copies will try
to delete the KmerAffectAnalyser.

This has been solved by using pointers to KmerSegmenter everywhere.