fuse.py: load() calls either load_vidjil() or load_clntab()
Previously, load_clntab() was launched when there was an exception raised by load_vidjil().
This was error prone, moreover we need the get the proper exception when there is a real error in load_vidjil().

Now the check is on the file name ('.clntab'). There could be better strategies, but
the detection of the file type has to be separated from the parsing itself.