test_browser.rb: renaming clone can work if we wait a little.
Because of 25a65b77, when focussing out of the input #new_name,
the whole model is updated within 500ms. But the previous test
(which renames using click) did focus out the input, therefore
the model is reupdated while we are performing the current test.

So do we really need to reupdate the whole model after 500ms when
focussing out the input? If yes, do we have a way to test if the
model has been updated (so that we can wait 'till it is updated).
That would be cleaner than waiting 1s at the beginning of the test.