WindowsStorage: getReads() can't return a reference to a list anymore.
Before the list was actually stored. Now the list of reads is not stored
anymore.  There are several lists of (some) reads, and for getReads() a single
list is computed from those lists.  We should see if it causes some slow down
but we rarely call getReads() and we don't have anymore to sample sequences.