vidjil.cpp: load incomplete germlines '-i' at the end, ignoring '-1'/'-2'
Loading incomplete germlines in the same index may give some problems,
as the same sequences will be loaded several times with different roles,
leading to AFFECT_AMBIGUOUS affectations. We thus ignore the '-1' setting
for incomplete germlines, and do not consider these germline for MAX12 ('-2').

Now '-i' should work with '-1' and/or '-2', and the previous test passes.