tests: remove TRD+ testing from trd-dd2-dd3.fa
TRDD2*01 is exactly 9 bp long. With SEED_9, we have in perfect Dd2 recombinations *one* matching k-mer.
The Dd2 TRD+ indexes have a load of around 0.05% and 0.16% (due to Dd3 or J), and thus, with the current e-value
computation, there is no chance to get a left e-value much smaller than 1e-3.