generate-recombinations: Generate V(D)J recombinations.
Contrarily to what the commit could let think, there is not so many modifications.
Basically the idea is to use the existing code and to transform generate_to_file
to a recursive function. So when generating a VDJ recombination we first do the
DJ recombination and then by calling the function recursively we use all the V
with the generated DJ.