Tests: Only check that we have the same number of Vs.
Do what the test says. We don't care if we have 143 affectations (it is lower now with -t
about 100). We want to see if we have the same affectations. Just do a sort -u:
if we have twice the same line the duplicate will be removed with the -u, hence we will have
only one occurrence of the XXX