Tests: Test that we only have the same segmentation for the first clone.
We changed several times this line to see if the clone is the same.
The purpose of the test is to check if we have the same clone in both datasets.
The purpose of the test is not to check the segmentation.

To check that, we remove the part of the lines that differ among both datasets
and only keep the segmentation information. Hence both lines should be equal and
should not appear in the diff.