vidjil.cpp: default window is now 50, both for VJ and V(D)J recombinations
Previously, the default -w was 40 for VJ, and 60 for V(D)J, except in
multi-locus identification (-g), when it was 40 even for V(D)J recombinations.
For the majority of cases, setting -w to 50 is thus more conservative.
Many users already tested -w 50 (or the less conservative -w 40) for IGH and
found correct results.

Now we have a uniform window size for all locus, allowing a better indexing
and comparison of windows across experiments and configurations.