windowExtractor.{h,cpp}: Stats are centralized in WindowExtractor.
Some statistics were stored in Germline while those statistics were filled in
WindowExtractor and WindowsStorage. Now they are stored and filled in
WindowExtractor.

We take benefit of this change to use BinReadStorage. It will give us more
detailed statistics on read lengths and clones/reads (the information is
there, but it is still not output yet).

By the way the function fillStatsClones is moved from WindowsStorage to WindowExtractor