tests: Update tests with new default -t option
Interestingly with -w 100, the number of foud windows is much larger
because now the window is better centered around the junction.
Before we could have spurious hits in the N-REGION which would
have shifted the window towards the 3' end.

Additionnally a few windows seem to be factorised.

(updated/merged by @magiraud)