Makefile: The rule must specify the output directory
A rule name is supposed to be the output file: if it doesn't exist the rule is relaunched.
To prevent relaunching every time compilation of vidjil and align.cgi we must provide
the output directory in the rule name so that Make knows where it should look for the output file.

By doing so the compilation will be relaunched only when the output file is older
than the files it depends on. We must also provide the library as a dependance for the
file to be rebuilt when the library changes.