tests/Makefile: runs should-vdj-tests also on revcomp'd files
All KmerSegmenter should-vdj tests (-2q) are now also run on revcomp'd files (-r), and
should give exactly the same output than the original test files.

Even if it works, we choose to not run FineSegmenter tests with the same option.
It would test nothing more, as the '-c segment' command first calls the KmerSegmenter to select
the best locus with its strand.