algo/: Remove delta_max, use delta_min only in FineSegmenter.
delta_max is now useless. It was use to make sure we didn't segment
something not robust enough. Now we use the e-value, which is a much
trustable value.

delta_min is useless to the KmerSegmenter (we can't have overlaps
with the first_max, last_max algorithms).
But it can still be useful to FineSegmenter, that's why we keep it.