core/germline.cpp: insert D k-mers in the index
There are sometimes common k-mers between D and V/J genes (this was at least
the case for IGH with -t 0), and this may give badly positioned windows.
Now all D k-mers are stored in the index, marking these k-mers as AMBIGUOUS.
Note that it is the first time that the D germline is actually used in the KmerSegmenter.