vidjil_utils: extract_fields_from_json can now take a partial JSON in input.
JSON files produced by Vidjil can be huge (dozens or hundreds of megabytes) and we
do not want to load them all. We just load a short prefix of the file.
Since the json module expects a well-formed JSON string we use the
cleanup_json_sample to close the JSON input