IKmerStore: lowest level insert is virtual
Since it can be redefined (and is redefined) we need to make the method
virtual so that the redefinition will be called even by some code
produced in IKmerStore.

We also need to tell in PointerACAutomaton that we must use the other
versions of insert that are defined in IKmerStore.