tests/Makefile: make shouldvdj_with_rc_merged
This merge all .should-vdj.fa files into one file, and test this file.

This test is twice faster than an usual 'make shouldvdj_with_rc': it should
be used to quickly check whether these tests run. However, when something fails,
it does not provide precisely the location of the error.