app.js: Change the way of requiring files
Load the files only when the previous ones have successfully loaded.
Apparently the require functions were called simultaneously and some files
(such as graph) that should be loaded by the fourth call to require was in
fact loaded before view (which is in the third call. Now the require functions
are called in the success callback function which ensures that previous files
have been properly loaded.