core/windowExtractor.cpp: do not halt when there is an error in a .fastq file
The .fastq parser in OnlineFasta() is rigourous. When the file was truncated somewhere,
there can be a non-valid sequence at the end of the file.

Previously Vidjil was halting in this case, and this could be quite frustating
when a large number of sequences were already processed. Now we just warn the user,
stop the analysis at this point, and properly output the clones.

Note that this does not affect the initial scan done on at most SAMPLE_APPROX_NB_SEQUENCES
sequences (currently 200): Vidjil will still halt on any error in these first sequences.