core/affectanalyser.cpp: better estimation of .max_found in .getMaximum(), more reads as UNSEG_AMBIGUOUS
Since the new flexible heuristic, introduced more than one year ago (de008a2, version 2014-07),
the idea behind the check of the segmentation point was:
  "Do we have enough affectations in good positions ('before' at the left and 'after' at the right) ?
   We tolerate some of them in bad positions, but there must be 'ratioMin' more in good positions."

However, the actual implementation of this idea was rather partial.
As a result, the 'ambiguous' sequence added in the previous commit was falsely segmented.

The new code is more symmetrical, .max_found being set to true when there are both:
 - more V at the left than V at the right, and than J at the left
 - more J at the right than J at the left, and than V at the right
"More" is defined by ratioMin and currently equals to 2.0.

A few reads that were previously segmented will now appear as UNSEG_AMBIGUOUS.