bug: When adding some kmers prevent the sequence being segmented
When we have few kmers at the start, the sequence will be segmented
(because first_pos_max is at the start of the sequence).

On the contrary if we add some kmers of the same locus further in the sequence
(in our case it is TGCTCCCCTA) the sequence won't be segmented because
first_pos_max is much larger and it is likely that having so few kmers of the
locus in such a large sequence is obtained by chance.

Clearly there is a problem with the first sequence that should not be
segmented: maybe the evalue should not be computed from position 1 to position
1+first_pos_max ?