germline: display the real information on germlines
The information on germlines were recovered through the index which does
not make sense with an Aho-Corasick automaton. In such a case the index
is the same but the germlines have different seeds and therefore index
loads should differ.

Since getIndexLoad depends on the KmerAffect, it also differs on V and J