file.py: upload message
One upload is finished, a message is displayed.
patient_id is a Row, so display the id instead of displaying something like Row(...).
Also request.vars['id'] is the file ID which doesn't mean anything to anyone, so replace
it with the filename.

Remove the semicolon at the end (we may have nothing more after that).