core/segment.cpp: CDR3 detection based on Cys104 and Phe118/Trp118 position
This was realized after prototyping and tests by @flothoni.
As in IMGT/JunctionAnalysis, the detection relies on the positions of Cys104 and Phe118/Trp118.
The detection is here in O(n), taking advantage of the already aligned V and J segments.

The current implementation will not give a precise positions when there are insertions or deletions
between Cys104 and the end of the V segment  (or between the start of J segment and Phe118/Trp118).
This could be improved by backtracking the DP matrix.