css: bikeshedding V, D, J main and highlight colors
The current colors were selected in d97cfefa6 in May 2014 and never changed.

This highly debatable commit sets new colors and new rendering for V/D/J segments,
with homogenous colors with the V/D/J segments highlighted when querying IMGT-VQUEST.
The goal is both to improve the contrast between the segments
and to be not too intrusive for the eye.