database.js initClones before loadGermline
Switched order of these two operations.
The reason being that initClones() is now responsible for calling a
conversion function on clone seg information in order to maintain
support for older files after the standard for seg definition has been
changed.
loadGermline() depends on seg information to determine which axis should
be created and its label.