css: border around the N
Use box-shadow instead of border as a box-shadow doesn't take any space (thanks to http://stackoverflow.com/a/18841365/1192742).
For some reason the shadow on the right cannot be seen so we put it on the left of the N, the D and the J (only when they are preceded by a N).