VidjilAuth.py Permissions Changes
In order to simplify the permissions system and improve performance as
well as make way for the new group associations. A single permission now
determines whether a group (or association of groups) has access to a
patient, run or sample_set

More advanced levels of access are now defined globally for a group.
Which means a group will always have the same permissions on files
accessible to it's association.

Furthermore groups are assigned in a parent/child manner. Only one group
can be a parent in an association, therefore associating a group to one
that already has a parent will associate said group to that parent.

A parent group will confer file readability to its children but the
children do not do the same for the parent and do not confer file
permissions with eachother.