graph.js: Clip the graph
When exporting the graph the gray area extended to the bottom.
We use a <clipPath> tag to ensure that nothing goes outside.
All the other elements are put in an element which refers to this clip path.
Moreover the rectangle defining the clip path is resized with the other rectangle
when the graph needs be resized.