getIndexLoad: Compute it on all the germlines
When the germline by itself didn't insert any kmer but relies
on other germlines (eg. for the XXX germline), we compute the
index load by relying on the index load of the other germlines.

Note that the sum may not accurately reflect the index load
(as there may have common k-mers between the germlines).