algo: Remove smallestAnalysableLength()
At first, we considered that smallestAnalysableLength() would be enough
to deal with differences between a k-mer index and an Aho-Corasick automaton.
This is actually not the case.

We need to have the length of the seed together with the affectation for the
Aho-Corasick automaton. In the end this allows to get the same affectation
strings than with a k-mer index. This would not have been possibble with the
smallesAnalysableLength() function and it makes it useless.