patient.py fix default selected group on create
This hotfix resolves an issue with default group selection when creating
a new patient.

This issue affected users that belonged to multiple groups but did not
have creation permissions on all groups.

The default user group is now also only displayed if the user has create
permissions on that group.