chimera.should-get: We segment again on IGL by chance
This is due to the Aho-Corasick automaton where we now have an index load for
the V gene and for the J gene. There is only one matching kmer from IGLJ but
that's enough to have an evalue < 1 because the index load for IGLJ is very
low and because there are only two reads in the dataset.

We could put a -e 1e-1 on the command line to make the test pass. Would this
be hacky or not?