tools.js: Totally rewrite computeStartStop()
The for-loop was useless. We start by defining the fields to be processed and then
for each of them we compute the start and stop positions.
We also take benefit of the fact that IMGT may define itself the start and stop positions
in which case we just retrieve them.