tools: Remove Ns at the end of a sequence
We remove the longest prefix (resp. suffix) ending with a
N (resp. starting) whose N content is greater or equal to
RATIO_TOO_MANY_N.

This is particularly useful for the representative: because of spaced
seeds we may have N in the representative, especially on the ends of the
representative (when only one seed matched at the extremity of the
sequence).