read_score.{h,cpp}: ReadQualityScorer
A new score depending on quality and length

The percentile of the quality is computed multiplied by the read length
and is then divided by what we consider being a good quality.
This somehow corresponds to artificially diminishing the read length if
the quality is bad.

We increase the number of bins to make sure that the better quality
reads will be in a different bin than medium quality reads.