data/no_representative.fa: Disambiguate a k-mer
Previously, by chance, AA.TA.CC.CC.CC had two occurrences.  This is a bit
cheating to modify this example for that reason but on the other hand, this is
just a small example and a small hit “by chance” like this one has greater
consequences. Yes I'm just rationalizing.

There is the same phenomenon at the end of the string.