core/representative: Do not rely anymore on k
k was the length of the contiguous seed. As we now have several seeds
this does not make sense anymore.

We now rely on the seed lengths.

Also, when extending the representative to the right we make sure that
the last position of the seed is on the position to be extended. This is
done to make sure that extending to the left or to the right are
equivalent (which was not the case in the previous implementation with
several seeds).