core/representative: The cover length is the most important.
What matters is how many positions are covered by seeds.
Therefore the main criteria to determine the best representative
is the highest cover length. In case of tie, the length of the representative
will be taken into account (is it really reasonable to take the longest one
in that case as we will take the noisest?)