models/task.py: run_copy takes name of the original filename
run_copy copied the sequence file into the result file. But the filename used was the internal Web2py filename (which is quite long).
And it was reencoded for the results_file which lead to a very long filename which was likely to exceed the limit of the
filename length.
Now we take the original filename for the data file, which will lead to a much shorter encoded filename