vidjil.cpp (tests and doc): -t options defaults again to 0
This reverts bf2f96fe and d317535.

After more than one year experience with -t 100 we realised that trimming leads to:
– missing special cases where the V gene is heavily trimmed
– giving bad unsegmentation causes: UNSEG only J appeared in cases where it didn't make sense
  those unsegmentation causes were replaced by UNSEG only V when changing -t to 0 (that also let us think that we could miss some V genes)

However -t 0 does have a drawback that should be addressed. It is more likely to have
spurious hits of V genes in the N-region (particularly with VDJ recombinations
where the N-region is longer). Those spurious hits tend to shift the window towards
the J gene.

In the tests the modifications are due either to slight changes in number of windows
(a small increase due to the previous reason) or to an increase in unsegmentation
with really large windows (since windows are shifted we may not have
enough space left to put a window).