algo/{vidjil, core/{windows, windowExtractor}}: Add sequences of interest
DNA sequences that should be found in the window (or windows that should
be found into it) can be kept whatever their concentration is. We
therefore extend what is already done for labelled windows.

Now the sequences do not need to be the length of the window to be found
but they can be of any length. However we require that the sequence of
interest contains the window (or is contained in it).