windows: read must be sampled depending on the score
The sample was constituted only on the sequence length which made sense
when the reads were stored depending on their length. But now the scoring
function changed and we focus on quality. It does not make sense anymore
to retrieve the longest reads then. We just want to have a sample of the
best reads (ie. those of better quality).

Using that, if we change our scoring function again, no modification
will be needed.

Note that the second parameter of getBestReads() is not used yet.
It may be useful to prevent too bad sequences to be sampled.

SequenceSampler is not used anymore. It will be removed in a future release.