kmerstore.h: satisfy clang
For an unknown reason, Clang complained about the line:
seed = IKmerStore<T>::seed
because “error: no viable overloaded '='”
It seemed to interpret one of the variable as const (which is not)
as it indicated:
/usr/bin/../lib/gcc/x86_64-linux-gnu/6.1.1/../../../../include/c++/6.1.1/bits/basic_string.h:565:7: note: candidate function not viable: 'this' argument has type 'const string' (aka 'const basic_string<char>'),
      but method is not marked const

The solution was to use a local variable. Clang was happy, but I don't
see what really makes the difference.