algo/test/Makefile: Comply with FreeBSD's sed
GNU sed and POSIX sed have slightly different behaviors.  Those behaviors are
incompatible if we don't specify a “backup” suffix with -i. So we specify such
a suffix and then we move all the files so that they overwrite the original
files.
Moreover in this case, the -e option is necessary for FreeBSD's sed.