algo/tests/Makefile: .bak files are the old files!
Never commit in a hurry!

The previous commit (see 8496cbb) indicated a suffix for the backup files
(necessary with POSIX sed). Then we were overwriting the tap files with the
bak ones. This was totally stupid as we were overwriting the correct files
with the old ones…

This is now fixed we just delete the backup files as we don't really need them
(we just specify the option to ensure the compatibility between POSIX and GNU
sed).