{segmenter,crossDomain}.js: Sequence IDs are now unique
Sequences sent to external website have now an unique identifier.
The clone index is prepended to the getName() (with a # as a separator).

This ensure that no sequence will share a same FASTA header, which could
cause some softwares to complain. Also, when analysing the results, we will
easily know which result belongs to which clone.

The clone index is put at the start rather than the end because
IMGT/V-QUEST cuts the header when it is too long.