vidjil.cpp: FineSegmenter in -c clones also considers e-value
Follows what is done for D in e6ffb91, and what was already done for V/J in -c segment.
Even if it may be counter-intuitive, the results of segmentation thus actually depends on the number of reads:
'-e 1' should mean that we expect one false positive segmentation in the output.
Even if some e-values slightly change, this commit does not affect the results of the tests.

We should investigate what happens when a FineSegmenter fails here.
Moreover, is it sane to have the same '-e' paramater for fixing both e-values of Kmer and FineSegmenter?
Or should we rather have another paramater?