controllers/default.py fix issue with init_db
There were two issue with the controller for ini_db:
Thre fist was the parameter passed to the function, causing web2py to
dismiss it as a controller.
The second was the truncate loop which was failing on foreign key
relationships. The truncate has been temporarily commented in order to
find a solution in the futur.