fuse.py: Change the criterion to decide information on which sample is chosen
Previously, the informations coming from the best top were kept.
However we could have a sequence in the top 1, with few reads, in a poorly segmented dataset that
was preferred over a top 2 sequence with lots of reads in a highly segmented dataset.

Now we just take into account the number of reads, by assuming with the higher number
of reads we should have the higher reliability. This should be a better criterion than the previous
one (wishful thinking).