modules/sample_set.py container object
Added a container object for sample_set objects.
This is in order to avoid using dictionnaries in the upcoming
controllers.
Although dictionnaries might serve the basic purpose, they lack to
ability to provide methods for data manipulation, and are therefore
rather limiting when it comes to providing polymorphism and abstraction.