models/sample_set_list.py extract list queries
The sample_set/all route contained mainy queries which not only gave it
a large amount of lines of code, but also made it difficult to test.
These changes place the queries in a class which represents a list of
sample sets.
Each querie has been isolated in its own method and can therefore be
called individually.
This will open the path for more thorough testing of our controllers and
also make them more maintainable and easier to understand.