docker-compose.yml remove vidjil volume mounts
Some volumes were added to mount vidjil from the host.
Although this makes it easier to handle the contents of the vidjil
directory, it's not very user-friendly and also means the installation
is not self-sufficient.
Removing the volumes lets the containers default to the vidjil
installation already contained within the vidjil docker image.