js/*.js: Fix error when db is undefined in multiple classes
The scatterplot, builder, graph, and list classes were not checking if
the database reference was defined, causing a JavaScript error when an
exception was thrown in the functions. This could only occur when not
using any database. This commit implements the behaviour seen in
segmenter.js: the exception is thrown if no database reference is defined
so that the exception can be seen in the JavaScript console, using a
function added to tools.js.