index.html, main.js start up indicator
The browser is quite large and can sometimes take some time to download
and start up.
Therfore the "patients" menu entry could sometimes take a long time to
be displayed if at all.
Adding this indicator will let the user know when everything has
finished loading and whether or not a server is configured with the
browser more clearly.