default.py: Timestamp with deleted file
Timestamp should not be ``deleted'' as a date is expected for the timestamp.
dateDiffMinMax can also deal with a "None" string. So we change the string
output here, which is less specific, but which works correctly with what is
expected in model.js.

The name of the sample is still "deleted" which will help remember that this
file has been deleted.

Fixes #2101.