file.py delete results when reuploading a file
If the results of a file a maintained after reuploading, any run on
another file would still contain the old version of the file in the
fuse, and file would be marked as deleted.

Not only is it confusing for the user to see a file marked as removed
where all files are seemingly present, but it is also not desirable to
have data that has been overwritten appear in a new analysis.